+++ catalyst/metrics/_ndcg.py
@@ -1,13 +1,10 @@
-from typing import Any, Dict, List
+from typing import Iterable
 
-import torch
-
-from catalyst.metrics._additive import AdditiveMetric
-from catalyst.metrics._metric import ICallbackBatchMetric
+from catalyst.metrics._topk_metric import TopKMetric
 from catalyst.metrics.functional._ndcg import ndcg
 
 
-class NDCGMetric(ICallbackBatchMetric):
+class NDCGMetric(TopKMetric):
     """
     Calculates the Normalized discounted cumulative gain (NDCG)
     score given model outputs and targets
@@ -49,10 +46,8 @@ class NDCGMetric(ICallbackBatchMetric):
         # {
         #     'ndcg01': 1.0,
         #     'ndcg02': 0.6131471991539001,
-        #     'ndcg': 1.0,
         #     'ndcg01/std': 0.0,
-        #     'ndcg02/std': 0.0,
-        #     'ndcg/std': 0.0
+        #     'ndcg02/std': 0.0
         # }
 
         metric.reset()
@@ -130,91 +125,20 @@ class NDCGMetric(ICallbackBatchMetric):
 
     def __init__(
         self,
-        topk_args: List[int] = None,
+        topk_args: Iterable[int] = None,
         compute_on_call: bool = True,
         prefix: str = None,
         suffix: str = None,
     ):
         """Init NDCGMetric"""
-        super().__init__(compute_on_call=compute_on_call, prefix=prefix, suffix=suffix)
-        self.metric_name_mean = f"{self.prefix}ndcg{self.suffix}"
-        self.metric_name_std = f"{self.prefix}ndcg{self.suffix}/std"
-        self.topk_args: List[int] = topk_args or [1]
-        self.additive_metrics: List[AdditiveMetric] = [
-            AdditiveMetric() for _ in range(len(self.topk_args))
-        ]
-
-    def reset(self) -> None:
-        """Reset all fields"""
-        for metric in self.additive_metrics:
-            metric.reset()
-
-    def update(self, logits: torch.Tensor, targets: torch.Tensor) -> List[float]:
-        """
-        Update metric value with ndcg for new data and return intermediate metrics values.
-
-        Args:
-            logits (torch.Tensor): tensor of logits
-            targets (torch.Tensor): tensor of targets
-
-        Returns:
-            list of ndcg@k values
-        """
-        values = ndcg(logits, targets, topk=self.topk_args)
-        values = [v.item() for v in values]
-        for value, metric in zip(values, self.additive_metrics):
-            metric.update(value, len(targets))
-        return values
-
-    def update_key_value(self, logits: torch.Tensor, targets: torch.Tensor) -> Dict[str, float]:
-        """
-        Update metric value with ndcg for new data and return intermediate metrics
-        values in key-value format.
-
-        Args:
-            logits (torch.Tensor): tensor of logits
-            targets (torch.Tensor): tensor of targets
-
-        Returns:
-            dict of ndcg@k values
-        """
-        values = self.update(logits=logits, targets=targets)
-        output = {
-            f"{self.prefix}ndcg{key:02d}{self.suffix}": value
-            for key, value in zip(self.topk_args, values)
-        }
-        output[self.metric_name_mean] = output[f"{self.prefix}ndcg01{self.suffix}"]
-        return output
-
-    def compute(self) -> Any:
-        """
-        Compute ndcg for all data
-
-        Returns:
-            list of mean values, list of std values
-        """
-        means, stds = zip(*(metric.compute() for metric in self.additive_metrics))
-        return means, stds
-
-    def compute_key_value(self) -> Dict[str, float]:
-        """
-        Compute ndcg for all data and return results in key-value format
-
-        Returns:
-            dict of metrics
-        """
-        means, stds = self.compute()
-        output_mean = {
-            f"{self.prefix}ndcg{key:02d}{self.suffix}": value
-            for key, value in zip(self.topk_args, means)
-        }
-        output_std = {
-            f"{self.prefix}ndcg{key:02d}{self.suffix}/std": value
-            for key, value in zip(self.topk_args, stds)
-        }
-        output_mean[self.metric_name_mean] = output_mean[f"{self.prefix}ndcg01{self.suffix}"]
-        output_std[self.metric_name_std] = output_std[f"{self.prefix}ndcg01{self.suffix}/std"]
-        return {**output_mean, **output_std}
+        super().__init__(
+            metric_name="ndcg",
+            metric_function=ndcg,
+            topk_args=topk_args,
+            compute_on_call=compute_on_call,
+            prefix=prefix,
+            suffix=suffix,
+        )
 
 
 __all__ = ["NDCGMetric"]